product single page fixed

+++ src/components/pages/crud/index.jsx
@@ -1,5 +1,6 @@
 import React, {Component} from 'react';
 import {Redirect} from "react-router-dom";
+import Swal from 'sweetalert2';
 
 import axios from 'axios';
 
@@ -18,12 +19,30 @@ class CrudIndex extends Component {
 
     delete = (category) =>{
 
-      axios.delete(`category/${category.id}`).then(() => {
-          
-        this.loadCategory();
-        // alert('Category deleted done :)');
+      Swal.fire({
+        title: 'Are you sure?',
+        text: "You won't be able to revert this!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, delete it!'
+      }).then((result) => {
+        if (result.isConfirmed) {
 
-     });
+          axios.delete(`category/${category.id}`).then(() => {
+            this.loadCategory();
+         });
+
+          Swal.fire(
+            'Deleted!',
+            'Your file has been deleted.',
+            'success'
+          )
+        }
+      })
+
+
         
       // let index = this.state.category.indexOf(category);
       // this.state.category.splice(index, 1);
@@ -54,9 +73,16 @@ class CrudIndex extends Component {
       
         axios.post(`category/update/${id}`, data).then(response =>{
 
-         this.loadCategory();
+          Swal.fire({
+            icon: 'success',
+            title: 'Category Updated'
+          })
+
+          this.loadCategory();
 
          this.setState({name: '', edit_category:false});
+
+         
 
         })
         .catch( e =>{
@@ -76,6 +102,23 @@ class CrudIndex extends Component {
 
          this.loadCategory();
          this.setState({name: ''});
+
+        const Toast = Swal.mixin({
+          toast: true,
+          position: 'top-end',
+          showConfirmButton: false,
+          timer: 3000,
+          timerProgressBar: true,
+          didOpen: (toast) => {
+            toast.addEventListener('mouseenter', Swal.stopTimer)
+            toast.addEventListener('mouseleave', Swal.resumeTimer)
+          }
+        })
+        
+        Toast.fire({
+          icon: 'success',
+          title: 'Category Added successfully'
+        })
 
         })
         .catch( e =>{
@@ -101,7 +144,6 @@ class CrudIndex extends Component {
    componentDidMount(){
 
      this.loadCategory();
-
     }
 
  render() {
